requirements.txt
# === CORE BIBLIOTECAS ===
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
filetype==1.2.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
orjson==3.10.18
packaging==24.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
tenacity==9.1.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
zstandard==0.23.0

# === GOOGLE & GENAI ===
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1

# === LANGCHAIN ECOSSISTEMA ===
langchain==0.3.26
langchain-core==0.3.67
langchain-google-genai==2.1.6
langchain-groq==0.3.5
langchain-text-splitters==0.3.8
langchain-community==0.3.27
langsmith==0.4.4
groq==0.29.0

# === FASTAPI E RELACIONADOS ===
fastapi==0.115.12
starlette==0.46.2
uvicorn==0.34.3

# === Pydantic e conexão com MySQL ===
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1

# === Alembic (Migrations) ===
Mako==1.3.10
MarkupSafe==3.0.2

# === Python-dotenv (ambiente) ===
python-dotenv==1.1.1
